requirements.txt
streamlit
transformers
torch
sentence-transformers
faiss-cpu
pandas
scikit-learn
requests
bert-score
beautifulsoup4
pyarrow